web/requirements.txt
requests
django-crispy-forms
